Crash fixes and adding Zenith Cluster item (#394)
* Add the new Zenith Clusters

* Explicitly cast Worship Totem Waves to ints

* Replace unsafe pow() uses with safer_math_pow()

* Standardized the big number to return when OverflowError occurs to a variable
* Moved safe data handling functions out of data_formatting and into new safer_data_handling file
* Moved a number of other functions out of data_formatting and into the consts file they're relevant to
* Capped the recorded level of Vault, Grimoire, Compass, and Tesseract bonuses to their expected max levels

* Fix logger imports

* Move mark_advice_completed to a method of the Advice class
