+++ mysite/caverns/glowshroom_tunnels.py
@@ -1,7 +1,6 @@
 from consts.progression_tiers import true_max_tiers
 from consts.consts_autoreview import EmojiType
 from models.models import AdviceSection, AdviceGroup, Advice
-from utils.data_formatting import mark_advice_completed
 from utils.logging import get_logger
 from flask import g as session_data
 # from consts.consts import glowshroom_tunnels_progressionTiers, break_you_best, ValueToMulti
@@ -358,7 +357,7 @@ def getJusticeAdviceGroup() -> AdviceGroup:
 
     for subgroup in cavern_advice:
         for advice in cavern_advice[subgroup]:
-            mark_advice_completed(advice)
+            advice.mark_advice_completed()
 
     cavern_ag = AdviceGroup(
         tier='',
