+++ mysite/caverns/shallow_caverns.py
@@ -3,7 +3,6 @@ from math import ceil
 from consts.progression_tiers import true_max_tiers
 from consts.consts_autoreview import EmojiType
 from models.models import AdviceSection, AdviceGroup, Advice
-from utils.data_formatting import mark_advice_completed
 from utils.logging import get_logger
 from flask import g as session_data
 # from consts.consts import shallow_caverns_progressionTiers, break_you_best, ValueToMulti
@@ -346,7 +345,7 @@ def getBraveryAdviceGroup(schematics) -> AdviceGroup:
 
     for subgroup in cavern_advice:
         for advice in cavern_advice[subgroup]:
-            mark_advice_completed(advice)
+            advice.mark_advice_completed()
 
     cavern_ag = AdviceGroup(
         tier='',
