+++ mysite/general/event_shop.py
@@ -1,5 +1,4 @@
 from models.models import AdviceSection, AdviceGroup, Advice
-from utils.data_formatting import mark_advice_completed
 from utils.logging import get_logger
 from flask import g as session_data
 
@@ -25,7 +24,7 @@ def getEventShopAdviceGroup() -> AdviceGroup:
     ))
     for advice in es_advice:
         if advice.resource == '':
-            mark_advice_completed(advice)
+            advice.mark_advice_completed()
 
     es_ag = AdviceGroup(
         tier='',
