+++ mysite/static/items.yaml
@@ -1986,6 +1986,7 @@ Quest106: Urie's Special Childhood Rock
 Quest107: Arcane Fragment # for the Arcane Cultist
 Quest108: Stamina Battery
 Quest109: Cropfall Basket
+Quest110: Zenith Cluster
 EquipmentSmithingTabs2: Anvil Tab II
 EquipmentSmithingTabs3: Anvil Tab III
 EquipmentSmithingTabs4: Anvil Tab IV
