+++ mysite/utils/number_formatting.py
@@ -1,9 +1,10 @@
+from consts.consts_autoreview import default_huge_number_replacement
 from utils.logging import get_logger
 
 logger = get_logger(__name__)
 
 
-def parse_number(number: str | int | float, default=1e100) -> int | float:
+def parse_number(number: str | int | float, default=default_huge_number_replacement) -> int | float:
     """
     Parses a string to a number. Simplifies to an int if possible. Otherwise, it returns a float
     """
