+++ mysite/w3/atom_collider.py
@@ -6,7 +6,6 @@ from consts.consts_w3 import buildings_tower_max_level, collider_storage_limit_l
 from consts.progression_tiers import atoms_progressionTiers
 from models.models import AdviceSection, AdviceGroup, Advice
 from utils.misc.add_subgroup_if_available_slot import add_subgroup_if_available_slot
-from utils.data_formatting import mark_advice_completed
 from utils.logging import get_logger
 from utils.text_formatting import pl
 
@@ -138,7 +137,7 @@ def getColliderSettingsAdviceGroup() -> AdviceGroup:
         ))
 
     for advice in settings_advice['Information']:
-        mark_advice_completed(advice)
+        advice.mark_advice_completed()
 
     settings_ag = AdviceGroup(
         tier='',
@@ -223,7 +222,7 @@ def getCostReductionAdviceGroup() -> AdviceGroup:
     ))
 
     for advice in cr_advice:
-        mark_advice_completed(advice)
+        advice.mark_advice_completed()
 
     cr_ag = AdviceGroup(
         tier='',
