+++ mysite/w3/death_note.py
@@ -1,6 +1,6 @@
 from math import floor
 from consts.consts_autoreview import break_you_best
-from consts.consts_general import current_world
+from consts.consts_idleon import current_world
 from consts.consts_w3 import apoc_names_list, apoc_difficulty_name_list, dn_delays, dn_skull_value_to_name_dict, getDNKillRequirement
 from consts.progression_tiers import deathNote_progressionTiers, true_max_tiers
 from flask import g as session_data
